% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frbs_api_key.R
\name{frbs_api_key}
\alias{frbs_api_key}
\title{Get the Firebase API key}
\usage{
frbs_api_key()
}
\value{
String
}
\description{
Gets your Firebase API key from the env var \code{FIREBASE_API_KEY}. Make sure
you have defined this variable in your .Renviron file.
}
\examples{
\dontrun{
frbs_api_key()
}
}
