% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frbs_error_handler.R
\name{frbs_error_handler}
\alias{frbs_error_handler}
\title{frbs error handler}
\usage{
frbs_error_handler(e)
}
\arguments{
\item{e}{The error.}
}
\description{
Used in the \code{tryCatch()} blocks of \verb{frbs_*()} functions
}
