% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frbs_send_password_reset_email.R
\name{frbs_send_password_reset_email}
\alias{frbs_send_password_reset_email}
\title{Send password reset email}
\usage{
frbs_send_password_reset_email(email)
}
\arguments{
\item{email}{User's email address.}
}
\value{
A named list with the following items:
\itemize{
\item \code{email}: Users' email address.
\item \code{error}: See Details.
}
}
\description{
Send password reset email
}
\details{
\itemize{
\item \code{error}:
\itemize{
\item \code{NULL} if no error code in response
\item A list of 2 if response was an error:
\itemize{
\item \code{code}: Error code
\item \code{message}: Error message
Visit \href{https://firebase.google.com/docs/reference/rest/auth#section-send-password-reset-email}{Firebase Auth REST API docs}
for more details
}
}
}
}
