% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCluster.Predict.R
\name{SBC.test}
\alias{SBC.test}
\title{SBC prediction phase}
\usage{
SBC.test(object, newdata)
}
\arguments{
\item{object}{the \code{\link{frbs-object}}.}

\item{newdata}{a matrix (\eqn{m \times n}) of data for the prediction process, where \eqn{m} is the number of instances and 
\eqn{n} is the number of input variables.}
}
\description{
This function is the internal function of the SBC method to compute the predicted values.
}
\seealso{
\code{\link{SBC}}
}
