\name{summary.frbs}
\alias{summary.frbs}
\title{The summary function for frbs objects}
\usage{
  \method{summary}{frbs} (object, ...)
}
\arguments{
  \item{object}{the \code{\link{frbs-object}}}

  \item{...}{the other parameters (not used)}
}
\description{
  This function enables the output of a summary of the
  \code{\link{frbs-object}}.
}

