% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frbs_change_password.R
\name{frbs_change_password}
\alias{frbs_change_password}
\title{Change password}
\usage{
frbs_change_password(id_token, password)
}
\arguments{
\item{id_token}{A Firebase Auth ID token for the user.}

\item{password}{User's new password.}
}
\value{
A named list with the following items:
\itemize{
\item \code{localId}:	The uid of the current user.
\item \code{email}:	User's email address.
\item \code{passwordHash}:	Hash version of password.
\item \code{providerUserInfo}:	A named list of of all linked provider objects which
contain "providerId" and "federatedId".
\item \code{idToken}:	New Firebase Auth ID token for user.
\item \code{refreshToken}:	A Firebase Auth refresh token.
\item \code{expiresIn}:	string	The number of seconds in which the ID token expires.
\item \code{error}:
\itemize{
\item \code{NULL} if no error code in response
\item A list of 2 if response was an error:
\itemize{
\item \code{code}: Error code
\item \code{message}: Error message
}
}
}
}
\description{
Change password
}
\details{
Visit \href{https://firebase.google.com/docs/reference/rest/auth#section-change-password}{Firebase Auth REST API docs}
for more details
}
\examples{
\dontrun{
# first sign in user and get the 'id_token':
user <- frbs_sign_in(email = "user@gmail.com", password = "password")
id_token <- user$idToken

# change password:
response <- frbs_change_password(
  id_token = id_token,
  password = "new-user-password"
)
response
}
}
