% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FRBS.MainFunction.R
\name{norm.data}
\alias{norm.data}
\title{The data normalization}
\usage{
norm.data(dt.ori, range.data, min.scale = 0, max.scale = 1)
}
\arguments{
\item{dt.ori}{a matrix (\eqn{n \times m}) of the original data.}

\item{range.data}{a matrix (\eqn{2 \times n}) containing the range of the data, where n is the number of variables, and
first and second rows are the minimum and maximum value, respectively.}

\item{min.scale}{the minimum value within normalization.}

\item{max.scale}{the maximum value within normalization.}
}
\value{
the normalized data
}
\description{
This function is to transform from real-valued data into normalized data.
}
\seealso{
\code{\link{denorm.data}}
}
